// src/app/api/models/absence.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AbsenceStatus } from '../models/absence-status';
import { AbsenceType } from '../models/absence-type';
import { User } from '../models/user';
export interface Absence {
  attachmentUrl?: string | null;
  endDate?: string;
  id?: string;
  startDate?: string;
  status?: AbsenceStatus;
  type?: AbsenceType;
  user?: User;
  userId?: string;
}